VIDEO-5167 | Fixing browser beta and unstanble tests
