+++ test/integration/index.js
@@ -3,9 +3,14 @@
 require('./spec/browserbugs/chromium');
 require('./spec/docker/reconnection');
 require('./spec/docker/docker');
+require('./spec/bandwidthprofile/regressions');
+require('./spec/bandwidthprofile/renderhints');
+require('./spec/bandwidthprofile/video');
 require('./spec/connect');
 require('./spec/logger');
-require('./spec/localparticipant');
+require('./spec/localparticipant/networkqualitylevel');
+require('./spec/localparticipant/publishunpublishtrack');
+require('./spec/localparticipant/regressions');
 require('./spec/localtracks');
 require('./spec/localtrackpublication');
 require('./spec/rest');
